pyproject.toml
[project]
name = "faithfulness-benchmark"
version = "0.1.0"
description = "AI Summarization Faithfulness Benchmark Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pypdf>=4.0.0",
    "requests>=2.31.0",
    "transformers>=4.36.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "nltk>=3.8.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
]

[project.scripts]
benchmark = "main:main"
